Made monster move method that randomly picks an atttack and uses it. Added a ton of help methods. Re implemented get_health in all enemy subclasses. It now takes a default argument numeric = False. changing it to true means that the method returns an integer reflecting the enemy/hero current health.
trying to build a for loop in zelda_battle function that
will make use of this change,
automatically ends the fight when an opponents health dips below 1.
Had it working a moment ago. I Suspect there was an
oversight with output streams and or return value

currently, the function outputs a single blank line
then returns the console to the user.